added random direction to playstate level transitions and modified state transitioning (this fixes a bug where low fps would make it almost impossible to transition state)

+++ src/entities/ui_canvas.hpp
@@ -47,7 +47,7 @@ public:
     }
 
     void GoDown() {
-        currentButtonID = currentButtonID + 1 < buttons.size() ? currentButtonID + 1 : 0;
+        currentButtonID = currentButtonID + 1 < (int)buttons.size() ? currentButtonID + 1 : 0;
     }
 
     void SetOffset(Vector2 newOffset) {
